hosted-agents/pure-python/simple-chat/requirements.txt
openai==1.82.0
azure-identity==1.25.3
starlette==0.45.3
uvicorn==0.34.3
